ASoC: WCD9304: Add the compander support to wcd9304
The compander system is intended to increase the dynamic
range of an input signal. But for the wcd9304 codec,
this feature is to use companding in a novel way in the DAC
converter path to increase the signal noise ratio.
The change is to add the compander support to wcd9304
codec. The dynamic compander enable/disable during stream
playback is not supported.

Change-Id: If912302ee9d48dde1318e894a982cba591908d8f
Signed-off-by: Kuirong Wang <kuirongw@codeaurora.org>
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>

Conflicts:
	include/linux/mfd/wcd9xxx/wcd9304_registers.h
	sound/soc/codecs/wcd9304-tables.c
	sound/soc/codecs/wcd9304.c