msm: camera: Invalidate the vpe_ctrl frame info field during close.
During VPE subdev close, if there is a pending frame with VPE HW,
we unmap the frame, but dont invalidate the frame info field. This
might cause issues if the VPE subdevice is opened again and closed
without sending a frame to VPE. Hence we need to invalidate this
field during close.

Change-Id: I8944af7b5233e277a42b15644e297332f9e38341
CRs-Fixed: 479655
Signed-off-by: Kiran Kumar H N <hurlisal@codeaurora.org>

Conflicts:
	drivers/media/video/msm/msm_vpe.c